requirements.txt
numpy==1.14.5
asgiref==3.3.1
certifi==2018.4.16
chardet==3.0.4
cloudpickle==0.5.3
coverage==4.5.1
cycler==0.10.0
Cython==0.29.22
dask==0.18.1
decorator==4.3.0
Django==3.1.6
django-compat==1.0.15
django-cors-headers==3.2.1
django-debug-toolbar==2.2
django-filter==2.4.0
django-webpack-loader==0.7.0
djangorestframework==3.11.2
ebaysdk==2.1.5
elasticsearch==5.5.3
gunicorn==19.9.0
idna==2.7
-e git://github.com/EdjoLabs/image-match.git#egg=image-match
kiwisolver==1.0.1
lxml==4.6.3
Markdown==2.6.11
matplotlib==2.2.2
networkx==2.1
Pillow==8.1.1
psycopg2-binary==2.8.6
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.4
PyWavelets==0.5.2
<<<<<<< HEAD
requests==2.23.0
scikit-image==0.12.3
=======
requests==2.25.1
scikit-image==0.14
>>>>>>> 073e460c62528fc39b2045f66d5f182ff08e4f33
scipy==1.1.0
six==1.14.0
sqlparse==0.3.1
toolz==0.9.0
urllib3==1.25.6
